Feature/npm3 inject

+++ scripts/postinstall.sh
@@ -2,8 +2,6 @@
 
 INFO="sails-mysql-transactions:"; # for console logs
 
-MOD_DIR="../../node_modules";
-
 set -e;
 
 # If this is an NPM installation, we do not expect `.gitmodules` in the directory
@@ -16,15 +14,15 @@ fi;
 
 # Check whether sails has been already installed or not. If not, this is an
 # error and we should not proceed.
-if [ ! -d "${MOD_DIR}/sails" ]; then
+if [ ! -d "../../node_modules/sails" ] && [ ! -d "../../node_modules/waterline" ]; then
     echo -e "\033[1;31m";
-    echo "${INFO} Sails installation not found!";
+    echo "${INFO} Sails and waterline installation not found!";
     echo "${INFO} Ensure your package.json, which has sails-mysql-transaction, also includes sails.";
     echo -e "\033[0m\n";
     exit 1;
 fi
 
-if [ -d "${MOD_DIR}/sails-mysql" ]; then
+if [ ! -d "../../node_modules/waterline" ] && [ -d "../../node_modules/sails-mysql" ]; then
     echo -e "\033[1;31m";
     echo "${INFO} WARNING - detected sails-mysql.";
     echo "${INFO} You may face unexpected behaviour.";
@@ -32,13 +30,29 @@ if [ -d "${MOD_DIR}/sails-mysql" ]; then
     echo -e "\033[0m\n";
 fi
 
-echo "${INFO} Injecting waterline...";
 
-pushd "${MOD_DIR}/sails" > /dev/null;
-npm remove waterline;
-npm install "${MOD_DIR}/sails-mysql-transactions/waterline";
-popd > /dev/null;
+if [ -d "../../node_modules/waterline" ]; then
+    echo "${INFO} Injecting waterline into sails...";
+    pushd "../../" > /dev/null;
+    npm remove waterline;
+    npm install "node_modules/sails-mysql-transactions/waterline";
+    popd > /dev/null;
+    
+    echo
+    echo "${INFO} Installation successful.";
+    echo
+    exit 0;
+fi
 
-echo
-echo "${INFO} Installation successful.";
-echo
+if [ -d "../../node_modules/sails" ]; then
+    echo "${INFO} Injecting waterline into sails...";
+    pushd "../../node_modules/sails" > /dev/null;
+    npm remove waterline;
+    npm install "../sails-mysql-transactions/waterline";
+    popd > /dev/null;
+    
+    echo
+    echo "${INFO} Installation successful.";
+    echo
+    exit 0;
+fi
